scheduler/requirements.txt
Django==1.11.7
django-autoslug==1.9.3
django-filter==1.1.0
django-taggit==0.22.1
djangorestframework==3.7.3
Markdown==2.6.9
pytz==2017.3
